requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# TI2BioP_nandy.py: 6
Pillow == 5.1.0

# Drawsp_qt.py: 2
# Prueba.py: 9
# TI2BioP_nandy.py: 5
# TIs4AAQt.py: 9
# locura.py: 7
# oTIsDialog.py: 10
# pondNAdialog.py: 4
# ponddialog.py: 4
# ponderation.py: 9
# ponderation_NA.py: 9
# reso_rc.py: 9
# ti2bioppQt.py: 9
PyQt5_sip == 12.7.0

# FastaProc.py: 8,9
# TI2BioP.py: 7
# TI2BioP_nandy.py: 4
# oTIsDialog.py: 5
biopython == 1.74

# FastaProc.py: 4,7
# TI2BioP.py: 12
matplotlib == 3.1.1

# FastaProc.py: 3
# TI2BioP_nandy.py: 3
# oTIsDialog.py: 6
# openct.py: 6
# openxfasta.py: 8
numpy == 1.17.4

# FastaProc.py: 14,15,16,17
# oTIsDialog.py: 13,14,15,16
propy == 1.0

# FastaProc.py: 5,6
scipy == 1.3.1
